requirements.txt
pathway>=0.13.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
